club/consumers.py
from channels.generic.websocket import AsyncWebsocketConsumer

class NotificationConsumer(AsyncWebsocketConsumer):
    async def connect(self):
        # Perform necessary authentication and authorization checks
        await self.accept()

    async def disconnect(self, close_code):
        # Clean up any resources related to the WebSocket connection
        pass

    async def receive(self, text_data):
        # Handle incoming WebSocket messages, if necessary
        pass

    async def send_notification(self, event):
        # Send a real-time notification to the WebSocket connection
        await self.send(text_data=event['message'])
