requirements.txt
asgiref==3.7.2
Django==4.1.1
djangorestframework_simplejwt==5.2.0
Requests==2.31.0
uvicorn==0.22.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
gunicorn==20.1.0
whitenoise==6.2.0
twilio==8.3.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
django-channels==0.7.0
